Added player horizontal movement collision
The player can now detect collisions while moving horizontally. It cannot detect vertical collisions yet. That will be next. So if you run the game in this state, you will have weird behavior when trying to land on top of a tile. It will move your player around very fast and out of the screen.