The player now has vertical collision detection
The player now detects collisions with tiles from all sides and for the most part it works. There are some issues that need to be ironed out with the collision detection because the player can be moving in both the x and y direction at the same time but it works enough for now.